close but no xh264
libxh264.so.160 isn't linking correctly at runtime in godot, but it's
very close to working now. libopenh264 attempts to encode, but seems
to fail despite producing a file.

pretty sue it's an issue with .so resolution within librecorder.so,
submitting this before I go too far off the beaten track while it works.

Also, added a simple thing for testing that librecorder.so works f2mp4.